+++ src/main/java/dev/easbarba/onur/database/Parse.java
@@ -21,8 +21,8 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.commons.io.FilenameUtils;
 
 import dev.easbarba.onur.domain.Configuration;
 import dev.easbarba.onur.domain.Project;
@@ -33,24 +33,26 @@ public final class Parse {
     public Parse() {
     }
 
+    // Parse one configuration file
     public Configuration one(File file) throws Exception {
         try {
             ObjectMapper mapper = new ObjectMapper();
             List<Project> project = Arrays.asList(mapper.readValue(file, Project[].class));
+            var name = FilenameUtils.removeExtension(file.getName());
 
-            return new Configuration(file.getName(), project);
+            return new Configuration(name, project);
         } catch (Exception ex) {
             throw ex;
         }
     }
 
-    // Parse all configuration files.
-    public Map<String, Configuration> all() {
-        Map<String, Configuration> configurations = new HashMap<String, Configuration>();
+    // Bundle all configuration files.
+    public List<Configuration> all() {
+        List<Configuration> configurations = new ArrayList<Configuration>();
 
         files.namespath().forEach(file -> {
             try {
-                configurations.put(file.getName(), one(file));
+                configurations.add(one(file));
             } catch (Exception ex) {
                 ex.printStackTrace();
             }