+++ src/main/java/dev/easbarba/onur/domain/Project.java
@@ -23,6 +23,7 @@ public final class Project {
 
     @JsonSetter(nulls = Nulls.SKIP)
     private String branch = "master";
+
     private String url;
 
     public Project() {